+++ src/Firebase/Auth/ApiClient.php
@@ -181,6 +181,14 @@ class ApiClient
         ]);
     }
 
+    public function verifyPassword(string $email, string $password): ResponseInterface
+    {
+        return $this->request('verifyPassword', [
+            'email' => $email,
+            'password' => $password,
+        ]);
+    }
+
     /**
      * @param string $idToken
      *